refactor(structure): migrate project files into src/ and update path handling :truck:
