+++ docker-compose.yml
@@ -7,3 +7,4 @@ services:
       - "3000:3000"
     environment:
       - NODE_ENV=production
+      - NODE_PATH=src
