+++ src/routes/general.js
@@ -1,8 +1,8 @@
 const router = require('express').Router()
-const allController = require('../controllers/all')
-const apiSignature = require('../middlewares/signature')
-const globalRules = require('../modules/global/rules')
-const { validateInput } = require('../validator/index')
+const allController = require('controllers/all')
+const apiSignature = require('middlewares/signature')
+const globalRules = require('modules/global/rules')
+const { validateInput } = require('utils/validators/index')
 
 /**
   * Example Routes
